preload-documents/requirements.txt
python-dotenv
openai
langchain
langchain-community
unstructured-client
pypdf2